Move S3DataSource to s3_helper
